flame graph support!
