+++ streamx-console/streamx-console-service/src/main/java/com/streamxhub/console/core/service/impl/ApplicationServiceImpl.java
@@ -490,6 +490,7 @@ public class ApplicationServiceImpl extends ServiceImpl<ApplicationMapper, Appli
         if (appParam.getFlameGraph()) {
             flameGraph = new HashMap<>();
             flameGraph.put("reporter", "com.streamxhub.plugin.profiling.reporters.HttpReporter");
+            flameGraph.put("type", ApplicationType.STREAMX_FLINK.getType());
             flameGraph.put("id", application.getId());
             flameGraph.put("url", properties.getConsoleUrl().concat("/flink/app/report"));
             flameGraph.put("token", Utils.uuid());
