+++ streamx-console/streamx-console-service/src/main/resources/application-dev.yml
@@ -54,7 +54,7 @@ streamx:
   max.batch.insert.num: 1000
   shiro:
     # 后端免认证接口 url
-    anonUrl: /passport/**,/systemName,/user/check/**,/websocket/**
+    anonUrl: /passport/**,/systemName,/user/check/**,/websocket/**,/metrics/**
     # token有效期，单位秒
     jwtTimeOut: 86400
   app-home: /Users/benjobs/Workspace/deploy/workspace
